January 2022 - Release
New diagram: Client Access Licenses
New site page: Product Name Changes
New diagram edit menu and features: Insert Image, Link, Notes
Please see the change log and user guide on m365maps.com for more information.

+++ js/saved.js
@@ -1,18 +1,11 @@
 /// <reference path="common.js" />
-/* global modalPrompt, setupOfflineIndicator, modalConfirm, getModalInputText,
-   isEmbedded, exportSvg, exportPng, StoreName, isIOS, setupModal */
+/* global showModalDialog, getModalInputText, isEmbedded, exportSvg, exportPng,
+   StoreName, isIOS, setupModal */
 
 /** Creates a visual divider for the item list. */
 function newDivider() {
   const element = document.createElement('div');
   element.className = 'divider';
-  return element;
-}
-
-/** Creates a container div to put the action buttons in. */
-function newActions() {
-  const element = document.createElement('div');
-  element.className = 'actions';
   return element;
 }
 
@@ -75,36 +68,47 @@ function addNewRow(container, key, entry) {
   const entryLink = newDiagramLink(entry.Title, `/viewsvg.htm#*${key}`);
   container.appendChild(entryLink);
 
-  const entryActions = newActions();
-  const thisDivider = newDivider();
-
   const entryRename = newRenameButton();
-  entryActions.appendChild(entryRename);
+  container.appendChild(entryRename);
+
+  const entryDelete = newDeleteButton();
+  container.appendChild(entryDelete);
+
+  const entrySvgButton = newSvgButton();
+  container.appendChild(entrySvgButton);
+
+  const entryPngButton = newPngButton();
+  container.appendChild(entryPngButton);
+
+  const thisDivider = newDivider();
+  container.appendChild(thisDivider);
+
   entryRename.addEventListener('click', function renameClick() {
-    modalPrompt('Rename saved diagram:', entryLink.textContent,
-      function renameOK() {
+    showModalDialog('Rename saved diagram:', true, entryLink.textContent,
+      'OK', function renameOK() {
         const newName = getModalInputText();
         if (!newName) return;
-
         entry.Title = newName;
         const entryJSON = JSON.stringify(entry);
         localStorage.setItem(key, entryJSON);
         entryLink.textContent = newName;
-      });
+      },
+      undefined, undefined,
+      'Cancel', undefined);
   });
 
-  const entryDelete = newDeleteButton();
-  entryActions.appendChild(entryDelete);
   entryDelete.addEventListener('click', function deleteClick() {
-    const message =
-      `Are you sure you want to delete "${entryLink.textContent}"?`;
-
-    modalConfirm(message,
-      function deleteYes() {
+    showModalDialog(
+      `Are you sure you want to delete "${entryLink.textContent}"?`,
+      false, undefined,
+      'Yes', function deleteYes() {
         localStorage.removeItem(key);
 
         container.removeChild(entryLink);
-        container.removeChild(entryActions);
+        container.removeChild(entryRename);
+        container.removeChild(entryDelete);
+        container.removeChild(entrySvgButton);
+        container.removeChild(entryPngButton);
         container.removeChild(thisDivider);
 
         if (container.getElementsByClassName('diagram').length === 0) {
@@ -113,11 +117,11 @@ function addNewRow(container, key, entry) {
           addEmptyMessage(container);
           container.appendChild(newDivider());
         }
-      });
+      },
+      undefined, undefined,
+      'No', undefined);
   });
 
-  const entrySvgButton = newSvgButton();
-  entryActions.appendChild(entrySvgButton);
   entrySvgButton.addEventListener('click',
     function svgLinkClick(event) {
       event.preventDefault();
@@ -125,8 +129,6 @@ function addNewRow(container, key, entry) {
     }
   );
 
-  const entryPngButton = newPngButton();
-  entryActions.appendChild(entryPngButton);
   entryPngButton.addEventListener('click',
     function pngLinkClick() {
       const background =
@@ -134,9 +136,6 @@ function addNewRow(container, key, entry) {
       exportPng(entry.Title + '.png', entry.SvgXml, background);
     }
   );
-
-  container.appendChild(entryActions);
-  container.appendChild(thisDivider);
 }
 
 /** Adds all the saved diagrams onto the page. */
@@ -149,7 +148,9 @@ function populateList() {
   let keys = [];
   for (let index = 0; index < localStorage.length; index += 1) {
     const key = localStorage.key(index);
-    if (key !== StoreName.Flags && key !== StoreName.Settings) {
+    if (key !== StoreName.Flags &&
+        key !== StoreName.Settings &&
+        key !== StoreName.MatrixSelection) {
       keys.push(key);
     }
   }
@@ -176,37 +177,54 @@ function populateList() {
   }
 }
 
+/** Imports the file at index from the files list. */
+function processImport(files, index) {
+  const file = files[index];
+
+  const reader = new FileReader();
+  reader.addEventListener('load',
+    function fileLoaded(event) {
+      showModalDialog(
+        files.length === 0 ?
+          'Import diagram as:' :
+          `Import diagram ${index + 1} of ${files.length} as:`,
+        true, file.name,
+        'OK', function importOK() {
+          const diagramTitle = getModalInputText();
+          if (!diagramTitle) return;
+
+          const storageKey = Date.now().toString();
+          const svgObject = {
+            Title: diagramTitle,
+            SvgXml: event.target.result,
+          };
+
+          const jsonData = JSON.stringify(svgObject);
+          localStorage.setItem(storageKey, jsonData);
+
+          populateList();
+
+          if (index !== files.length - 1) {
+            index++;
+            processImport(files, index);
+          }
+        },
+        undefined, undefined,
+        files.length === 1 ? 'Cancel' : 'Skip',
+        function importSkipOrCancel() {
+          if (index !== files.length - 1) {
+            processImport(files, ++index);
+          }
+        });
+    }
+  );
+
+  reader.readAsText(file);
+}
+
 /** Imports the selected files into the Saved Diagrams list. */
 function filesSelected(event) {
-  const reader = new FileReader();
-
-  for (const file of event.target.files) {
-    reader.addEventListener('load',
-      function fileLoaded(event) {
-        modalPrompt('Import diagram as:', file.name,
-          function importOK() {
-            const diagramTitle = getModalInputText();
-            if (!diagramTitle) return;
-
-            const storageKey = Date.now().toString();
-            const svgObject = {
-              Title: diagramTitle,
-              SvgXml: event.target.result,
-            };
-
-            const jsonData = JSON.stringify(svgObject);
-            localStorage.setItem(storageKey, jsonData);
-
-            const container = document.getElementById('collection');
-            addNewRow(container, storageKey, svgObject);
-          });
-      }
-    );
-
-    reader.readAsText(file);
-  }
-
-  event.target.value = '';
+  processImport(event.target.files, 0);
 }
 
 /** Clicks the export link corresponding with the supplied exportType for
@@ -226,7 +244,6 @@ function DOMContentLoaded() {
     document.getElementById('menu').style.display = 'none';
   }
 
-  setupOfflineIndicator();
   setupModal();
   populateList();
